# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for tbls database schema analysis.\n\n`tbls serve` starts an HTTP server that provides endpoints to analyze databases\nand return schema information as JSON. Supports asynchronous task execution with\nprogress polling, checkpoint/resume, and task cancellation.",
        "title": "tbls serve API",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://github.com/k1LoW/tbls/blob/main/LICENSE"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/schema": {
            "post": {
                "description": "Analyze a database asynchronously. Returns a task ID immediately for progress polling.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schema"
                ],
                "summary": "Submit schema analysis task",
                "parameters": [
                    {
                        "description": "Schema analysis request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cmd.SchemaRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Task accepted",
                        "schema": {
                            "$ref": "#/definitions/cmd.TaskAcceptedResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/cmd.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/schema/status/{task_id}": {
            "get": {
                "description": "Get the status and progress of a schema analysis task.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schema"
                ],
                "summary": "Get task status",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Task ID",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task status",
                        "schema": {
                            "$ref": "#/definitions/cmd.TaskStatusResponse"
                        }
                    },
                    "404": {
                        "description": "Task not found",
                        "schema": {
                            "$ref": "#/definitions/cmd.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/schema/{task_id}": {
            "delete": {
                "description": "Cancel a running task. The current progress is saved to a checkpoint file.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schema"
                ],
                "summary": "Cancel task",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Task ID",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task cancelled",
                        "schema": {
                            "$ref": "#/definitions/cmd.TaskCancelledResponse"
                        }
                    },
                    "400": {
                        "description": "Task is not running",
                        "schema": {
                            "$ref": "#/definitions/cmd.TaskNotRunningError"
                        }
                    },
                    "404": {
                        "description": "Task not found",
                        "schema": {
                            "$ref": "#/definitions/cmd.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "cmd.CheckpointConfig": {
            "description": "Checkpoint/resume configuration",
            "type": "object",
            "properties": {
                "enabled": {
                    "description": "Enabled enables checkpoint/resume functionality",
                    "type": "boolean"
                },
                "force": {
                    "description": "Force ignores existing checkpoint and starts fresh",
                    "type": "boolean"
                },
                "ttl": {
                    "description": "TTL is the checkpoint validity duration (Go duration format)",
                    "type": "string",
                    "example": "24h"
                }
            }
        },
        "cmd.CommentConfig": {
            "description": "Table and column comments configuration",
            "type": "object",
            "properties": {
                "columnComments": {
                    "description": "ColumnComments maps column names to their comments",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "table": {
                    "description": "Table is the table name",
                    "type": "string",
                    "example": "users"
                },
                "tableComment": {
                    "description": "TableComment is the comment for the table",
                    "type": "string",
                    "example": "User accounts table"
                }
            }
        },
        "cmd.DSNConfig": {
            "description": "Database connection configuration",
            "type": "object",
            "required": [
                "url"
            ],
            "properties": {
                "headers": {
                    "description": "Headers are custom headers for HTTP-based connections",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "url": {
                    "description": "URL is the database connection string (required)",
                    "type": "string",
                    "example": "postgres://user:pass@localhost:5432/dbname"
                }
            }
        },
        "cmd.DetectVirtualRelationsConfig": {
            "description": "Virtual relation detection configuration",
            "type": "object",
            "properties": {
                "enabled": {
                    "description": "Enabled enables virtual relation detection",
                    "type": "boolean"
                },
                "strategy": {
                    "description": "Strategy is the naming strategy (default, rails, laravel)",
                    "type": "string",
                    "enum": [
                        "default",
                        "rails",
                        "laravel"
                    ]
                }
            }
        },
        "cmd.ErrorResponse": {
            "description": "Generic error response",
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error is the error message",
                    "type": "string",
                    "example": "dsn.url is required"
                }
            }
        },
        "cmd.FormatConfig": {
            "description": "Output format settings",
            "type": "object",
            "properties": {
                "adjust": {
                    "description": "Adjust column widths",
                    "type": "boolean"
                },
                "sort": {
                    "description": "Sort tables and columns alphabetically",
                    "type": "boolean"
                }
            }
        },
        "cmd.ProgressInfo": {
            "description": "Detailed progress information during stats collection",
            "type": "object",
            "properties": {
                "completed_columns": {
                    "description": "CompletedColumns is the number of columns completed",
                    "type": "integer",
                    "example": 45
                },
                "current_column": {
                    "description": "CurrentColumn is the currently processing column",
                    "type": "string",
                    "example": "customer_id"
                },
                "current_table": {
                    "description": "CurrentTable is the currently processing table",
                    "type": "string",
                    "example": "orders"
                },
                "total_columns": {
                    "description": "TotalColumns is the total number of columns to process",
                    "type": "integer",
                    "example": 200
                }
            }
        },
        "cmd.RelationConfig": {
            "description": "Table relation configuration",
            "type": "object",
            "properties": {
                "cardinality": {
                    "description": "Cardinality describes the relationship cardinality",
                    "type": "string",
                    "enum": [
                        "zero or one",
                        "exactly one",
                        "zero or more",
                        "one or more"
                    ]
                },
                "columns": {
                    "description": "Columns are the child table columns",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "author_id"
                    ]
                },
                "def": {
                    "description": "Def is the relation definition/description",
                    "type": "string"
                },
                "parentColumns": {
                    "description": "ParentColumns are the parent table columns",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "id"
                    ]
                },
                "parentTable": {
                    "description": "ParentTable is the parent table name",
                    "type": "string",
                    "example": "users"
                },
                "table": {
                    "description": "Table is the child table name",
                    "type": "string",
                    "example": "posts"
                }
            }
        },
        "cmd.SchemaRequest": {
            "description": "Request body for submitting a schema analysis task",
            "type": "object",
            "required": [
                "dsn"
            ],
            "properties": {
                "comments": {
                    "description": "Comments defines additional comments to add",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cmd.CommentConfig"
                    }
                },
                "desc": {
                    "description": "Desc is the database description",
                    "type": "string"
                },
                "detectVirtualRelations": {
                    "description": "DetectVirtualRelations configures virtual relation detection",
                    "allOf": [
                        {
                            "$ref": "#/definitions/cmd.DetectVirtualRelationsConfig"
                        }
                    ]
                },
                "distance": {
                    "description": "Distance is the relation distance for filtering",
                    "type": "integer"
                },
                "dsn": {
                    "description": "DSN contains the database connection configuration",
                    "allOf": [
                        {
                            "$ref": "#/definitions/cmd.DSNConfig"
                        }
                    ]
                },
                "exclude": {
                    "description": "Exclude specifies tables to exclude (supports wildcards)",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "*_backup",
                        "*_temp",
                        "*_log"
                    ]
                },
                "force": {
                    "description": "Force ignores checkpoint and forces stats collection",
                    "type": "boolean"
                },
                "format": {
                    "description": "Format contains output format settings",
                    "allOf": [
                        {
                            "$ref": "#/definitions/cmd.FormatConfig"
                        }
                    ]
                },
                "include": {
                    "description": "Include specifies tables to include (supports wildcards)",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "users",
                        "posts",
                        "comments_*"
                    ]
                },
                "name": {
                    "description": "Name overrides the database name",
                    "type": "string"
                },
                "relations": {
                    "description": "Relations defines additional relations to add",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/cmd.RelationConfig"
                    }
                },
                "stats": {
                    "description": "Stats configures statistics collection",
                    "allOf": [
                        {
                            "$ref": "#/definitions/cmd.StatsConfig"
                        }
                    ]
                }
            }
        },
        "cmd.StatsConfig": {
            "description": "Statistics collection configuration",
            "type": "object",
            "properties": {
                "checkpoint": {
                    "description": "Checkpoint configures checkpoint/resume functionality",
                    "allOf": [
                        {
                            "$ref": "#/definitions/cmd.CheckpointConfig"
                        }
                    ]
                },
                "enabled": {
                    "description": "Enabled enables statistics collection",
                    "type": "boolean"
                },
                "exclude": {
                    "description": "Exclude specifies tables to exclude from stats collection",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "include": {
                    "description": "Include specifies tables to collect stats for (supports wildcards)",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "largeTableThreshold": {
                    "description": "LargeTableThreshold is the row count threshold for large table sampling",
                    "type": "integer",
                    "example": 1000000
                },
                "recentDays": {
                    "description": "RecentDays is days to look back for large table sampling",
                    "type": "integer",
                    "example": 30
                },
                "sampleSize": {
                    "description": "SampleSize is the maximum rows to sample",
                    "type": "integer",
                    "example": 10000
                },
                "topN": {
                    "description": "TopN is the number of top values to collect per column",
                    "type": "integer",
                    "example": 10
                }
            }
        },
        "cmd.TaskAcceptedResponse": {
            "description": "Response when a schema analysis task is accepted",
            "type": "object",
            "properties": {
                "status": {
                    "description": "Status is the initial task status",
                    "type": "string",
                    "example": "pending"
                },
                "task_id": {
                    "description": "TaskID is the unique task identifier for polling",
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "cmd.TaskCancelledResponse": {
            "description": "Response when a task is cancelled",
            "type": "object",
            "properties": {
                "checkpoint_saved": {
                    "description": "CheckpointSaved indicates if progress was saved to checkpoint",
                    "type": "boolean",
                    "example": true
                },
                "status": {
                    "description": "Status is the task status after cancellation",
                    "type": "string",
                    "example": "cancelled"
                },
                "task_id": {
                    "description": "TaskID is the task identifier",
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "cmd.TaskNotRunningError": {
            "description": "Error response when task is not running",
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error is the error message",
                    "type": "string",
                    "example": "task is not running"
                },
                "status": {
                    "description": "Status is the current task status",
                    "type": "string",
                    "example": "completed"
                },
                "task_id": {
                    "description": "TaskID is the task identifier",
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "cmd.TaskStatusResponse": {
            "description": "Task status and progress information",
            "type": "object",
            "properties": {
                "completed_at": {
                    "description": "CompletedAt is the task completion time",
                    "type": "string",
                    "example": "2025-01-04T10:35:00Z"
                },
                "error": {
                    "description": "Error contains the error message if failed",
                    "type": "string",
                    "example": "connection refused"
                },
                "progress": {
                    "description": "Progress contains detailed progress information",
                    "allOf": [
                        {
                            "$ref": "#/definitions/cmd.ProgressInfo"
                        }
                    ]
                },
                "result": {
                    "description": "Result contains the schema analysis result",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_k1LoW_tbls_schema.Schema"
                        }
                    ]
                },
                "resumed_from_checkpoint": {
                    "description": "ResumedFromCheckpoint indicates if task was resumed from checkpoint",
                    "type": "boolean"
                },
                "stage": {
                    "description": "Stage is the current processing stage",
                    "enum": [
                        "analyzing",
                        "collecting_stats",
                        "inferring",
                        "completed",
                        "failed",
                        "cancelled"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/stats.Stage"
                        }
                    ],
                    "example": "collecting_stats"
                },
                "started_at": {
                    "description": "StartedAt is the task start time",
                    "type": "string",
                    "example": "2025-01-04T10:30:00Z"
                },
                "status": {
                    "description": "Status is the current task status (pending, running, completed, failed, cancelled)",
                    "type": "string",
                    "enum": [
                        "pending",
                        "running",
                        "completed",
                        "failed",
                        "cancelled"
                    ],
                    "example": "running"
                },
                "task_id": {
                    "description": "TaskID is the task identifier",
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "dict.Dict": {
            "type": "object"
        },
...
        "github_com_k1LoW_tbls_schema.Schema": {
            "type": "object",
            "properties": {
                "desc": {
                    "type": "string"
                },
                "driver": {
                    "$ref": "#/definitions/schema.Driver"
                },
                "enums": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.Enum"
                    }
                },
                "functions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.Function"
                    }
                },
                "inferences": {
                    "$ref": "#/definitions/schema.SchemaInferences"
                },
                "labels": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.Label"
                    }
                },
                "name": {
                    "type": "string"
                },
                "relations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.Relation"
                    }
                },
                "tables": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.Table"
                    }
                },
                "viewpoints": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.Viewpoint"
                    }
                }
            }
        },
...
        "schema.Driver": {
            "type": "object",
            "properties": {
                "database_version": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/schema.DriverMeta"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "schema.DriverMeta": {
            "type": "object",
            "properties": {
                "current_schema": {
                    "type": "string"
                },
                "dict": {
                    "$ref": "#/definitions/dict.Dict"
                },
                "search_paths": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "schema.Enum": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "values": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "schema.Function": {
            "type": "object",
            "properties": {
                "arguments": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "return_type": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "schema.Label": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "virtual": {
                    "type": "boolean"
                }
            }
        },
...
        "schema.SchemaInferences": {
            "type": "object",
            "properties": {
                "business_insights": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.BusinessInsight"
                    }
                },
                "foreign_keys": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/schema.InferredForeignKey"
                    }
                }
            }
        },
...
        "stats.Stage": {
            "type": "string",
            "enum": [
                "analyzing",
                "collecting_stats",
                "inferring",
                "completed",
                "failed",
                "cancelled"
            ],
            "x-enum-varnames": [
                "StageAnalyzing",
                "StageCollectingStats",
                "StageInferring",
                "StageCompleted",
                "StageFailed",
                "StageCancelled"
            ]
        }
    },
...
}